weston: Do not override PACKAGECONFIG
PACKAGECONFIG should not be overridden. If not, it will become very complex
to enable other features defined in the main recipe.

Also, removes the duplicate PACKAGECONFIG settings defined in the
bbappend after this change.

For sololite, remove egl from PACKAGECONFIG as egl is enabled by default for
DISTRO_FEATURES wayland.

This changed is already upstreamed to community.

Signed-off-by: Neena Busireddy <neena.busireddy@freescale.com>